+++ eva/executor/exchange_executor.py
@@ -18,12 +18,12 @@ from ray.util.queue import Queue
 
 from eva.executor.abstract_executor import AbstractExecutor
 from eva.executor.executor_utils import ExecutorError
-from eva.experimental.ray.executor.ray_stage import (
+from eva.executor.helpers.ray_stage import (
     StageCompleteSignal,
     ray_stage,
     ray_stage_wait_and_alert,
 )
-from eva.experimental.ray.planner.exchange_plan import ExchangePlan
+from eva.plan_nodes.exchange_plan import ExchangePlan
 from eva.models.storage.batch import Batch
