+++ test/optimizer/test_optimizer_task.py
@@ -18,7 +18,7 @@ from unittest.mock import patch
 from mock import MagicMock
 
 from eva.configuration.configuration_manager import ConfigurationManager
-from eva.experimental.ray.planner.exchange_plan import ExchangePlan
+from eva.plan_nodes.exchange_plan import ExchangePlan
 from eva.optimizer.cost_model import CostModel
 from eva.optimizer.operators import (
     LogicalFilter,
